Upgrade build and codebase to C++20
- Set CMAKE_CXX_STANDARD to 20 in all CMakeLists.txt files.
- Replace manual workarounds with C++20 features (std::isnan, std::isfinite, std::all_of, std::remove_cvref_t).
- Switch mm::source_location to use std::source_location where available.
- Add support for std::u8string_view in AnsiOstream.
- Implement transparent lookup for emoji HexPrefixTree using std::unordered_map.
- Update CRTP base classes in Crtp.h with public constructors to support aggregate initialization.
- Fix build regressions for MSVC, macOS, and Wasm introduced by the C++20 standard upgrade.

+++ src/global/emojis.cpp
@@ -612,7 +612,7 @@ NODISCARD QString mmqt::encodeEmojiShortCodes(const QString &s)
 
             // REVISIT: should this include "U+" or not?
             char hex[32];
-            snprintf(hex, sizeof(hex), "[:U+%X:]", c);
+            snprintf(hex, sizeof(hex), "[:U+%X:]", static_cast<unsigned int>(c));
             for (const char ascii : std::string_view{hex}) {
                 m_output += static_cast<char32_t>(static_cast<uint8_t>(ascii));
             }
